move party files to parties folder
